pyproject.toml
[project]
name = "cloudcli-server-kubernetes"
version = "0.1.0"
description = "Cloudcli Server Kubernetes"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[librabbitmq,sqlalchemy]>=5.5.1",
    "click>=8.1.8",
    "fastapi>=0.115.12",
    "flower>=2.0.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "retry>=0.9.2",
    "ruamel-yaml>=0.18.10",
    "uvicorn[standard]>=0.34.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "celery[pytest]>=5.5.1",
    "pytest>=8.3.5",
]

[project.scripts]
cloudclik8s = "cloudcli_server_kubernetes.cli:main"
